Configure npm to save exact versions
`npm install --save <pkg>` and `npm install --save-dev <pkg>`
save dependencies into package.json with a semver range by default.

This isn't safe because there's no guarantee that a new version of
a package won't introduce breaking changes.

Instead of using semver ranges, the `--save-exact` flag should be used
to specify exact versions of dependencies in the package.json file.

To make it even easier, define `save-exact=true` in a .npmrc file so
that `--save-exact` is automatically set whenever `npm install` is used
to save a package into the project.